Recup bdd boutique + Modif class admin

+++ class/admin.php
@@ -21,18 +21,15 @@ class admin
         $query_ajoutcategorie = mysqli_query($this->get('bdd')->getco(), "INSERT INTO categorie (nom,description) VALUE ('" . $newcat . "','" . $description . "')");
         $this->bdd->checkcategorie();
 
-        foreach ($this->bdd->get('categorie') as $cat)
-        {
-            if($cat['nom'] == $newcat)
-            {
+        foreach ($this->bdd->get('categorie') as $cat) {
+            if ($cat['nom'] == $newcat) {
                 $idnewcat = $cat['id'];
             }
         }
 
         $this->traitementimgcat($idnewcat);
 
-        if($this->filesimg != NULL)
-        {
+        if ($this->filesimg != NULL) {
             $query_ajoutcategorie_img = mysqli_query($this->get('bdd')->getco(), "UPDATE categorie SET img = '" . $this->filesimg . "' WHERE id = '" . $idnewcat . "'");
         }
     }
@@ -54,7 +51,10 @@ class admin
         $id_cat = $_GET['idcat'];
         $recup_img_delete = mysqli_query($this->get('bdd')->getco(), "SELECT img FROM categorie WHERE id = '$id_cat'");
         $result_img = mysqli_fetch_assoc($recup_img_delete);
-        unlink('../'.$result_img['img']);
+
+        if (basename($result_img['img']) != 'default.png') {
+            unlink('../' . $result_img['img']);
+        }
 
         $query_deletecategorie = mysqli_query($this->get('bdd')->getco(), "DELETE FROM categorie WHERE id = '$id_cat'");
     }
@@ -66,11 +66,9 @@ class admin
             $imageFileType = strtolower(pathinfo(basename($_FILES["img"]["name"]), PATHINFO_EXTENSION));
             $target_dir = "img/cat/";
             $target_file = $target_dir . $id . '.' . $imageFileType;
-            foreach (scandir($target_dir) as $files)
-            {
-                if(pathinfo(basename($files), PATHINFO_FILENAME) == $id)
-                {
-                    unlink($target_dir.$files);
+            foreach (scandir($target_dir) as $files) {
+                if (pathinfo(basename($files), PATHINFO_FILENAME) == $id) {
+                    unlink($target_dir . $files);
                 }
             }
             $check = getimagesize($_FILES["img"]["tmp_name"]);
@@ -90,16 +88,14 @@ class admin
 
     public function ajoutsouscategorie($newsouscat, $description, $cat)
     {
-        $recupidcat = mysqli_query($this->get('bdd')->getco(), "SELECT id FROM categorie WHERE nom = '".$cat."'");
+        $recupidcat = mysqli_query($this->get('bdd')->getco(), "SELECT id FROM categorie WHERE nom = '" . $cat . "'");
         $idcat = mysqli_fetch_row($recupidcat);
 
         $query_ajoutsouscategorie = mysqli_query($this->get('bdd')->getco(), "INSERT INTO sous_categorie (nom,description,id_categorie) VALUE ('" . $newsouscat . "','" . $description . "','" . $idcat[0] . "')");
         $this->bdd->checksouscategorie();
 
-        foreach ($this->bdd->get('souscategorie') as $souscat)
-        {
-            if($souscat['nom'] == $newsouscat)
-            {
+        foreach ($this->bdd->get('souscategorie') as $souscat) {
+            if ($souscat['nom'] == $newsouscat) {
                 $idnewsouscat = $souscat['id'];
                 echo $idnewsouscat;
             }
@@ -107,15 +103,14 @@ class admin
 
         $this->traitementimgsouscat($idnewsouscat);
 
-        if($this->filesimg != NULL)
-        {
+        if ($this->filesimg != NULL) {
             $query_ajoutsouscategorie_img = mysqli_query($this->get('bdd')->getco(), "UPDATE sous_categorie SET img = '" . $this->filesimg . "' WHERE id = '" . $idnewsouscat . "'");
         }
     }
 
     public function majsouscategorie($majsouscat, $majdescription, $idcat, $idsouscat)
     {
-        $recupidcat = mysqli_query($this->get('bdd')->getco(), "SELECT id FROM categorie WHERE nom = '".$idcat."'");
+        $recupidcat = mysqli_query($this->get('bdd')->getco(), "SELECT id FROM categorie WHERE nom = '" . $idcat . "'");
         $idcat = mysqli_fetch_row($recupidcat);
 
         $this->traitementimgsouscat($idsouscat);
@@ -132,7 +127,7 @@ class admin
         $id_souscat = $_GET['idsouscat'];
         $recup_img_delete = mysqli_query($this->get('bdd')->getco(), "SELECT img FROM sous_categorie WHERE id = '$id_souscat'");
         $result_img = mysqli_fetch_assoc($recup_img_delete);
-        unlink('../'.$result_img['img']);
+        unlink('../' . $result_img['img']);
 
         $query_deletesouscategorie = mysqli_query($this->get('bdd')->getco(), "DELETE FROM sous_categorie WHERE id = '$id_souscat'");
     }
@@ -144,11 +139,9 @@ class admin
             $imageFileType = strtolower(pathinfo(basename($_FILES["img"]["name"]), PATHINFO_EXTENSION));
             $target_dir = "img/sous_cat/";
             $target_file = $target_dir . $id . '.' . $imageFileType;
-            foreach (scandir($target_dir) as $files)
-            {
-                if(pathinfo(basename($files), PATHINFO_FILENAME) == $id)
-                {
-                    unlink($target_dir.$files);
+            foreach (scandir($target_dir) as $files) {
+                if (pathinfo(basename($files), PATHINFO_FILENAME) == $id) {
+                    unlink($target_dir . $files);
                 }
             }
             $check = getimagesize($_FILES["img"]["tmp_name"]);
